refactor!: reorg storage

+++ contracts/test/TestReentrancyAttack.sol
@@ -5,8 +5,8 @@ import "../protocol/IRigoblockV3Pool.sol";
 
 contract TestReentrancyAttack {
     address private immutable RIGOBLOCK_POOL;
-    uint public count = 0;
-    uint private maxLoopCount = 2;
+    uint256 public count = 0;
+    uint256 private maxLoopCount = 2;
 
     constructor(address rigoblockPool) {
         RIGOBLOCK_POOL = rigoblockPool;
@@ -19,9 +19,9 @@ contract TestReentrancyAttack {
     function mintPool() public {
         count += 1;
         if (count <= maxLoopCount) {
-            try IRigoblockV3Pool(payable(RIGOBLOCK_POOL)).mint(address(this), 1e19, 1) {
-
-            } catch Error(string memory reason) {
+            try IRigoblockV3Pool(payable(RIGOBLOCK_POOL)).mint(address(this), 1e19, 1) {} catch Error(
+                string memory reason
+            ) {
                 revert(reason);
             }
         }
